Account for top padding when adding the columns
Same problem found in data2imgX1.m

With the previous code each column was drawn (values set to 1) top to bottom starting in px and increasing until bar(I).

The starting position was correct, but the ending position was wrong as it didn't account for the column being displaced downwards 'px' pixels.